+++ src/main/java/insane96mcp/survivalreimagined/module/mining/inventory/ForgeResultSlot.java
@@ -48,7 +48,6 @@ public class ForgeResultSlot extends Slot {
         this.checkTakeAchievements(pStack);
     }
 
-    //TODO Award experience for forging
     protected void checkTakeAchievements(ItemStack pStack) {
         pStack.onCraftedBy(this.player.level, this.player, this.removeCount);
         if (this.player instanceof ServerPlayer && this.container instanceof ForgeBlockEntity forgeBlockEntity) {